reshuffle and start day 9, need to restructure it.

+++ src/int_code/mod.rs
@@ -3,8 +3,8 @@ use std::io::{self};
 
 pub mod machine;
 
-pub fn read_file(path: &str) -> io::Result<Vec<i32>> {
+pub fn read_file(path: &str) -> io::Result<Vec<i64>> {
     Ok(fs::read_to_string(path)?.split(',')
-        .map(|x| x.parse::<i32>().unwrap())
+        .map(|x| x.parse::<i64>().unwrap())
         .collect())
 }
