+++ src/permutation/mod.rs
@@ -1,10 +1,10 @@
-fn _perm2(input: Vec<i32>) ->  Vec<Vec<i32>> {
+fn _perm2(input: Vec<i64>) ->  Vec<Vec<i64>> {
     let f = input[0];
     let s = input[1];
     vec![vec![f.clone(), s.clone()], vec![s, f]]
 }
 
-fn _recursive_perm(mut input: Vec<i32>) -> Vec<Vec<i32>> {
+fn _recursive_perm(mut input: Vec<i64>) -> Vec<Vec<i64>> {
     match input.len() {
         0 => Vec::new(),
         1 => vec![input],
@@ -15,7 +15,7 @@ fn _recursive_perm(mut input: Vec<i32>) -> Vec<Vec<i32>> {
             let mut sub_perms = _recursive_perm(input);
             while let Some(sub_perm) = sub_perms.pop() {
                 for pos in 0..=sub_perm.len() {
-                    let mut perm: Vec<i32> = sub_perm.clone();
+                    let mut perm: Vec<i64> = sub_perm.clone();
                     perm.insert(pos, first.clone());
                     permutations.push(perm);
                 }
@@ -25,6 +25,6 @@ fn _recursive_perm(mut input: Vec<i32>) -> Vec<Vec<i32>> {
     }
 }
 
-pub fn permutations(input: Vec<i32>) -> Vec<Vec<i32>> {
+pub fn permutations(input: Vec<i64>) -> Vec<Vec<i64>> {
     _recursive_perm(input)
 }
